whisper/requirements.txt
flask==2.3.2
whisper==1.0.0
werkzeug==2.3.7
python-dotenv==1.0.0